CLN: Standardize udf files across backends
