+++ ibis/backends/postgres/udf.py
@@ -24,12 +24,12 @@ class PostgresUDFError(IbisError):
     pass
 
 
-def ibis_to_pg_sa_type(ibis_type):
+def _ibis_to_pg_sa_type(ibis_type):
     """Map an ibis DataType to a Postgres-compatible sqlalchemy type"""
     return to_sqla_type(ibis_type, type_map=PostgreSQLExprTranslator._type_map)
 
 
-def sa_type_to_postgres_str(sa_type):
+def _sa_type_to_postgres_str(sa_type):
     """Map a Postgres-compatible sqlalchemy type to a Postgres-appropriate
     string"""
     if callable(sa_type):
@@ -37,12 +37,12 @@ def sa_type_to_postgres_str(sa_type):
     return sa_type.compile(dialect=sa_postgres_dialect())
 
 
-def ibis_to_postgres_str(ibis_type):
+def _ibis_to_postgres_str(ibis_type):
     """Map an ibis DataType to a Postgres-appropriate string"""
-    return sa_type_to_postgres_str(ibis_to_pg_sa_type(ibis_type))
+    return _sa_type_to_postgres_str(_ibis_to_pg_sa_type(ibis_type))
 
 
-def create_udf_node(name, fields):
+def _create_udf_node(name, fields):
     """Create a new UDF node type.
 
     Parameters
@@ -108,7 +108,7 @@ def existing_udf(name, input_types, output_type, schema=None, parameters=None):
     )
     udf_node_fields['resolve_name'] = lambda self: name
 
-    udf_node = create_udf_node(name, udf_node_fields)
+    udf_node = _create_udf_node(name, udf_node_fields)
 
     def _translate_udf(t, expr):
         func_obj = sa.func
@@ -175,10 +175,10 @@ $$;
 """
 
     postgres_signature = ', '.join(
-        '{name} {type}'.format(name=name, type=ibis_to_postgres_str(type_))
+        '{name} {type}'.format(name=name, type=_ibis_to_postgres_str(type_))
         for name, type_ in zip(parameter_names, in_types)
     )
-    return_type = ibis_to_postgres_str(out_type)
+    return_type = _ibis_to_postgres_str(out_type)
     # If function definition is indented extra,
     # Postgres UDF will fail with indentation error.
     func_definition = dedent(inspect.getsource(python_func))
