--- ibis/backends/postgres/udf/__init__.py
@@ -1,3 +1,0 @@
-from ibis.backends.postgres.udf.api import existing_udf, udf
-
-__all__ = ('existing_udf', 'udf')
